+++ src/implementations/types.ts
@@ -1,5 +1,4 @@
 export enum ScopeTypes {
-  BANKING = 'BANKING',
   CHECKDIGIT = 'CHECKDIGIT',
   EXPERIMENTAL = 'EXPERIMENTAL'
 }